computerVision/faceRecognition/requirements.txt
opencv-python
face_recognition
Pillow
pymongo
python-dotenv
requests
numpy
